pyproject.toml
[project]
name = "aillm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai==1.12.1",
    "google-generativeai>=0.8.5",
    "python-dotenv==1.1.0",
]
